backend/app/crawler/parsers/utils.py
import time
from functools import wraps
from io import BytesIO

from PIL import Image


def retry(exception, tries=3, delay=2, backoff=2, logger=None):
    def deco_retry(f):

        @wraps(f)
        def f_retry(*args, **kwargs):
            mtries, mdelay = tries, delay
            while mtries > 1:
                try:
                    return f(*args, **kwargs)
                except exception as e:
                    msg = f"{str(e)}, Retrying in {mdelay} seconds..."
                    if logger:
                        logger.warning(msg)
                    else:
                        print(msg)
                    time.sleep(mdelay)
                    mtries -= 1
                    mdelay *= backoff

            return f(*args, **kwargs)

        return f_retry  # true decorator

    return deco_retry


@retry(OSError)
def get_pil_image_from_response_content(content):
    file = BytesIO(content)
    img = Image.open(file)
    return img
