backend/requirements/app.txt
Django==2.1.2
django-rest-framework==0.1.0
djangorestframework==3.8.2
pytz==2018.5
setuptools==40.4.3
psycopg2==2.7.4
wheel==0.32.0
uwsgi==2.0.17
Pillow==5.3.0
django-cors-headers==2.4.0
redis==2.10.6
celery==4.2.1
boto3==1.9.57
requests==2.19.1
beautifulsoup4==4.6.0
lxml==4.2.1
drf-nested-routers==0.91
elasticsearch-dsl==6.3.1
httmock==1.2.6